Refactored, adjusted the gradients and condensed the contributors to take up fewer rows when the window gets expanded
