+++ package.json
@@ -4,10 +4,11 @@
   "private": true,
   "dependencies": {
     "framer-motion": "^1.3.0",
-    "react": "^16.8.6",
+    "mui": "^0.0.1",
+    "react": "^16.14.0",
     "react-dom": "^16.8.6",
     "react-github-btn": "^1.0.5",
-    "react-scripts": "3.0.1",
+    "react-scripts": "^5.0.0",
     "styled-components": "^4.3.2"
   },
   "scripts": {
