+++ src/assets/DeveloperCard.js
@@ -1,29 +1,34 @@
 import styled from "styled-components";
 // styling for the developer card in footer
 const DeveloperCard = styled.div`
-  display: inline-block;
-  width: 10em;
-  height: 10em;
-  margin: 3em;
-  flex: 0 0 15%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: center;
+  width: 8em;
+  height: 8e,;
+  margin: 2.25em;
 
   .dev-pic {
     margin-bottom: 1em;
     border-radius: 100%;
-    width: 10em;
-    max-height:10em;
+    width: 8em;
+    height: 8em;
   }
   .dev-name {
+    color: #dde;
     line-height: 1.7em;
-    font-weight: 100;
-  }
-  a {
-    color: white;
-    text-decoration: none;
     font-weight: 500;
   }
+  a {
+    color: #dde;
+    text-decoration: none;
+    font-size: .9em;
+    font-weight: 500;
+    margin-bottom: -2.25em;
+  }
   a:hover {
-    color: #3fd1eb;
+    color: #eef;
     transition: 0.3s;
   }
 `;
