+++ src/assets/styles.css
@@ -1,5 +1,5 @@
 /* macbook air */
-@import url('https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.9.0/css/all.min.css');
+@import url("https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.9.0/css/all.min.css");
 
 @media only screen and (max-height: 766px) {
   .header {
@@ -66,11 +66,11 @@ html {
   margin-top: 3rem;
 }
 
-.App {
+#app {
   text-align: center;
   align-content: center;
   justify-content: center;
-  font-family: 'Quattrocento Sans', sans-serif;
+  font-family: "Quattrocento Sans", sans-serif;
   color: white;
 }
 
@@ -101,7 +101,6 @@ html {
   align-items: center;
 }
 
-
 .lowerText {
   font-size: 1.5rem;
   display: block;
@@ -114,16 +113,18 @@ html {
 }
 
 .footer {
-  background: rgb(37, 96, 159);
-  background: linear-gradient(107deg,
-      rgba(37, 96, 159, 1) 0%,
-      rgba(20, 37, 54, 1) 100%);
+  background: rgb(1, 3, 131);
+  background: linear-gradient(
+    107deg,
+    rgba(37, 160, 100, 1) 0%,
+    rgba(20, 55, 35, 1) 100%
+  );
   padding: 1em;
 }
 
-.devCards{
+.devCards {
   display: flex;
-  flex-flow: row;
+  flex-direction: row;
   flex-wrap: wrap;
   justify-content: center;
 }
@@ -155,7 +156,7 @@ html {
   height: 1000px;
   box-shadow: 0px 8px 15px 5px rgba(66, 66, 66, 0.8);
   background: black;
-  -webkit-transform:scale(0.85);
+  -webkit-transform: scale(0.85);
   -moz-transform: scale(0.85);
 }
