+++ src/components/Contribute.js
@@ -57,8 +57,8 @@ export const Contribute = () => {
     <BottomDiv>
       <h2>Want to Contribute?</h2>
       <div>OverVue is open source. Help make Vue prototyping better!</div>
-      <a class="contribute" href="https://github.com/open-source-labs/OverVue">
-        <p>Github </p> <i class="fab fa-github" />
+      <a className="contribute" href="https://github.com/open-source-labs/OverVue">
+        <p>Github </p> <i className="fab fa-github" />
       </a>
     </BottomDiv>
   );
